NewbieViewModel을 Protocol로 수정
- NewbiewViewModelProtocol을 따르는 CreateGroupCodeViewModel, CreateGroupViewModel, UserCodeViewModel 생성
